// src/app/models/page.ts
export class Page {
    id = -1;
    title: string = '';
    shop_id: number;
    handle: string = '';
    body_html: string = '';
    author: string = '';
    published: number;
    published_at: string = '';
    template_suffix: string = '';
    metafields_global_title_tag: string = '';
    metafields_global_description_tag: string = '';
    constructor() {
    }
}